requirement.txt
streamlit
googlesearch-python
pandas
nltk
